+++ RELEASE-NOTES.md
@@ -4,6 +4,11 @@
 
 This pre-version of the Angular-Wakanda Pack is based on alpha releases of Angular-Wakanda which notes are available in the next section.
 
+###v0.2.6
+
+* update to angular-wakanda version 0.0.11
+* remove alternative code required for previous versions
+
 ###v0.2.5
 
 * update to angular-wakanda version 0.0.10
@@ -15,26 +20,37 @@ Understand this module is still under development and the current releases are a
 
 The versions provided can be unstable, the features may not be finished. This document notes may not be completly accurate.
 
+###v0.0.11
+
+* added $fetch, $toJSON, $isLoaded on nested collection
+* integrated patch on error between null and $_deferred (undefined)
+
 ###v0.0.10
+
 * fixed another "Converting circular structure to JSON" bug in .$toJSON() method
 
 ###v0.0.9
+
 * fixed "Converting circular structure to JSON" bug in .$toJSON() method
 
 ###v0.0.8
+
 * $fetch on deferred entities (not yet on collections)
 * $isLoaded method (makes it easier to check if your entity/collection was fetched or not)
 * user defined collection method at the root of your collection (not on the nested ones)
 * added .$toJSON() on entities and root collections (not yet on the nested ones)
 
 ###v0.0.7
+
 * added calculated attributes
 * 1>n relationships (no deffered, no collection methods)
 
 ###v0.0.6
+
 * bug fix on undefined object
 
 ###v0.0.5
+
 * added photo src retrieving support
 
 ###...